Initial Docs (#398)
* initial antora + content

* fix copypasta

* iPad multitasking

* Resize and optimise all images

This change is required to ensure that the images are as small as
possible and that their dimensions are consistent (also that they're not
massive in size and obscure the rest of the content).

* Update images

The existing images were optimised and additional images (also 
optimised) were added.

* Enable Antora’s experimental options

This is so that the button, menu, and keyboard binding highlights can
be used to draw more attention to the relevant text.

* Update existing content

A majority of the existing information’s been updated, whether that’s 
refining the grammar, spelling, or use of AsciiDoc formatting. 
Additionally, new content’s been added, bringing the documentation 
closer to completion.

* Updated the navigation file

* Update the spelling of the component name

* Update the SEO information in the documentation

* Further update the app documentation

This change makes a series of small fixes, whether to grammar, spelling,
or flow. It also notably extends the settings documentation.

* Update the Quick Access documentation

* Update the SEO metadata in index, settings, and troubleshooting

* Revised accounts, collaboration, files integration, mdm, & security

* Minor updates to the iOS collaboration, integration, & files docs

* Update the collaboration, files integration, & settings docs

* Finalise the settings documentation

This change completes the view and revoke previously approved
certificates sections.

* Update docs/modules/ROOT/pages/ios_collaboration.adoc

Co-Authored-By: Jesús Recio Rincón <jrecio@solidgear.es>

* Update docs/modules/ROOT/pages/ios_settings.adoc

~~Always~~ ==> Immediately

Co-Authored-By: Jesús Recio Rincón <jrecio@solidgear.es>

* Planned Logging Feature: Mask private Data

* Updated theme names

* Resolve public-/private-link confusion

* Update OAuth2 implementation for iOS 12

* Updating ios_security.adoc with latest changes from SECURITY.md

Updates from aa2bdac8beff19a22dfc3bddc8a22d867e66ac9e